+++ lets/modules/encode/powershell/obfuscate/compress.py
@@ -9,9 +9,6 @@ class Compress(DockerExtension, Module):
     """
     Obfuscate a powershell script with compression-based manipulation.
     """
-    TECHNIQUES = [
-        "1",    # Convert entire command to one-liner and compress
-    ]
 
     def usage(self) -> object:
         """
@@ -21,13 +18,6 @@ class Compress(DockerExtension, Module):
         :return: ArgumentParser object
         """
         parser = super().usage()
-
-        # Specify technique
-        parser.add_argument("-t", "--technique",
-            help="technique to use in obfuscation",
-            type=str,
-            choices=self.TECHNIQUES,
-            default="1")
 
         return parser
 
@@ -48,13 +38,11 @@ class Compress(DockerExtension, Module):
         # Validate input
         try:
             assert data, "Expecting data"
-            technique = self.options.get("technique")
-            assert technique in self.TECHNIQUES, "Invalid technique: %s (%s)" % (technique, str(self.TECHNIQUES))
         except AssertionError as e:
             self.throw(e)
 
         # Build command
-        cmd = "Invoke-Obfuscation -Quiet -ScriptPath /data/in -Command 'COMPRESS,%(technique)s' | Out-File /data/out" % self.options
+        cmd = "Invoke-Obfuscation -Quiet -ScriptPath /data/in -Command 'COMPRESS,1' | Out-File /data/out" % self.options
 
         # Prepare input and output files
         with self.IO(data,
@@ -84,39 +72,38 @@ class Compress(DockerExtension, Module):
         Perform unit tests to verify this module's functionality.
         """
         # Test encoding
-        for technique in self.TECHNIQUES:
-            test = string.ascii_letters + string.digits
-            testcmd = "Write-Output '%s'" % test
-            encoded = b"".join(self.do(testcmd.encode(), {"technique" : technique}))
-            self.assertTrue(len(encoded) > 0, "Technique %s produced inaccurate results" % technique)
+        test = string.ascii_letters + string.digits
+        testcmd = "Write-Output '%s'" % test
+        encoded = b"".join(self.do(testcmd.encode()))
+        self.assertTrue(len(encoded) > 0, "Encoding printable characters produced inaccurate results")
 
-            # Try to evade errors with missing variables
+        # Try to evade errors with missing variables
+        compare = encoded.decode().lower()
+        while (
+            "pshome" in compare or
+            "comspec" in compare or
+            "shellid" in compare or
+            "verbosepreference" in compare or
+            "mdr" in compare
+            ):
+            encoded = b"".join(self.do(testcmd.encode()))
             compare = encoded.decode().lower()
-            while (
-                "pshome" in compare or
-                "comspec" in compare or
-                "shellid" in compare or
-                "verbosepreference" in compare or
-                "mdr" in compare
-                ):
-                encoded = b"".join(self.do(testcmd.encode(), {"technique" : technique}))
-                compare = encoded.decode().lower()
 
-            # Base64 encode to get past the container layer
-            cmd = base64.b64encode(encoded.decode().encode("utf-16-le")).decode()
+        # Base64 encode to get past the container layer
+        cmd = base64.b64encode(encoded.decode().encode("utf-16-le")).decode()
 
-            # Test execution
-            with self.Container(
-                image="local/tools/invoke-obfuscation:latest",
-                network_disabled=True,
-                entrypoint=["pwsh", "-e"],
-                command=[cmd]) as container:
+        # Test execution
+        with self.Container(
+            image="local/tools/invoke-obfuscation:latest",
+            network_disabled=True,
+            entrypoint=["pwsh", "-e"],
+            command=[cmd]) as container:
 
-                # Fetch output
-                output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]
+            # Fetch output
+            output = [line.strip() for line in container.logs(stdout=True, stderr=True)][0]
 
-                # Wait for container to cleanup
-                container.wait()
+            # Wait for container to cleanup
+            container.wait()
 
-                # Verify execution was successful
-                self.assertEqual(output.decode(), test, "Execution of technique %s produced inaccurate results" % technique)
+            # Verify execution was successful
+            self.assertEqual(output.decode(), test, "Execution produced inaccurate results")